split unit tests, fix adapters path
